+++ src/Constants.ts
@@ -225,11 +225,18 @@ export const CONSTANTS: IMap<any> = {
     TotalNumBitNodes: 24,
 
     LatestUpdate: `
-    v0.52.2 - 2021-07-10 TITLE (hydroflame & community)
+    v0.52.2 - 2021-07-15 Oh yeah, BN11 is a thing (drunk hydroflame tbh)
     -------------------------------------------
 
+    ** Source-Files **
+
+    * Source-File 11 now also provides a small reduction to the price increase
+      multiplier.
+
     ** Augmentations **
+
     * New Augmentation offered by Aevum, themed around 777 and offers some basic
       programs.
+    * Augmentation descriptions are now more concise and consistent.
 `,
 }